+++ apps/blog-app/vite.config.ts
@@ -6,11 +6,15 @@ import { defineConfig } from 'vite';
 // https://vitejs.dev/config/
 export default defineConfig(() => {
   return {
+    root: __dirname,
     publicDir: 'src/assets',
     optimizeDeps: {
       include: ['@angular/common'],
     },
     build: {
+      outDir: '../../dist/apps/blog-app/client',
+      reportCompressedSize: true,
+      commonjsOptions: { transformMixedEsModules: true },
       target: ['es2020'],
     },
     plugins: [