+++ apps/trpc-app/vite.config.ts
@@ -8,6 +8,7 @@ import { nxViteTsPaths } from '@nx/vite/plugins/nx-tsconfig-paths.plugin';
 // https://vitejs.dev/config/
 export default defineConfig(({ mode }) => {
   return {
+    root: __dirname,
     publicDir: 'src/public',
     optimizeDeps: {
       include: ['@angular/common', '@angular/forms', 'isomorphic-fetch'],
@@ -16,6 +17,9 @@ export default defineConfig(({ mode }) => {
       noExternal: ['@analogjs/trpc', '@trpc/server'],
     },
     build: {
+      outDir: '../../dist/apps/trpc-app/client',
+      reportCompressedSize: true,
+      commonjsOptions: { transformMixedEsModules: true },
       target: ['es2020'],
     },
     plugins: [
@@ -33,6 +37,11 @@ export default defineConfig(({ mode }) => {
       splitVendorChunkPlugin(),
     ],
     test: {
+      reporters: ['default'],
+      coverage: {
+        reportsDirectory: '../../coverage/apps/trpc-app',
+        provider: 'v8',
+      },
       globals: true,
       environment: 'jsdom',
       setupFiles: ['src/test-setup.ts'],