+++ libs/card/vite.config.ts
@@ -6,8 +6,10 @@ import angular from '@analogjs/vite-plugin-angular';
 // https://vitejs.dev/config/
 export default defineConfig(({ mode }) => {
   return {
+    root: __dirname,
     plugins: [angular()],
     test: {
+      reporters: ['default'],
       globals: true,
       environment: 'jsdom',
       setupFiles: ['src/test-setup.ts'],