+++ packages/nx-plugin/src/generators/app/generator.spec.ts
@@ -14,7 +14,7 @@ import { AnalogNxApplicationGeneratorOptions } from './schema';
 describe('nx-plugin generator', () => {
   const setup = async (
     options: AnalogNxApplicationGeneratorOptions,
-    nxVersion = '17.0.0',
+    nxVersion = '18.0.0',
     standalone = false
   ) => {
     const tree = createTreeWithEmptyWorkspace(
@@ -120,6 +120,36 @@ describe('nx-plugin generator', () => {
     expect(devDependencies['vitest']).toBe('^0.32.2');
   };
 
+  const verifyCoreDependenciesNxV18_AngularV17_X = (
+    dependencies: Record<string, string>,
+    devDependencies: Record<string, string>
+  ) => {
+    expect(dependencies['@analogjs/content']).toBe('^0.2.0');
+    expect(dependencies['@analogjs/router']).toBe('^0.2.0');
+    expect(dependencies['@angular/platform-server']).toBe(
+      dependencies['@angular/core']
+    );
+    expect(dependencies['front-matter']).toBe('^4.0.2');
+    expect(dependencies['marked']).toBe('^5.0.2');
+    expect(dependencies['marked-gfm-heading-id']).toBe('^3.1.0');
+    expect(dependencies['marked-highlight']).toBe('^2.0.1');
+    expect(dependencies['mermaid']).toBe('^10.2.4');
+    expect(dependencies['prismjs']).toBe('^1.29.0');
+
+    expect(devDependencies['@nx/devkit']).toBe('^18.0.0');
+    expect(devDependencies['@nx/angular']).toBe('^18.0.0');
+    // we just check for truthy because @nx/eslint generator
+    // will install the correct version based on Nx version
+    expect(devDependencies['@nx/eslint']).toBeTruthy();
+    expect(devDependencies['@analogjs/platform']).toBe('^0.2.0');
+    expect(devDependencies['@analogjs/vite-plugin-angular']).toBe('^0.2.0');
+    expect(devDependencies['@nx/vite']).toBe('^18.0.0');
+    expect(devDependencies['jsdom']).toBe('^22.1.0');
+    expect(devDependencies['vite']).toBe('~5.0.0');
+    expect(devDependencies['vite-tsconfig-paths']).toBe('^4.2.0');
+    expect(devDependencies['vitest']).toBe('^1.0.4');
+  };
+
   const verifyConfig = (
     config: ProjectConfiguration,
     name: string,
@@ -168,8 +198,7 @@ describe('nx-plugin generator', () => {
     config: ProjectConfiguration,
     devDependencies: Record<string, string>
   ) => {
-    expect(config.targets.lint).toBeDefined();
-    expect(devDependencies['eslint']).toBeDefined();
+    expect(devDependencies['@nx/eslint']).toBeDefined();
   };
 
   const verifyTailwindIsSetUp = (
@@ -235,10 +264,70 @@ describe('nx-plugin generator', () => {
     expect(injectsTrpcClient).toBeFalsy();
   };
 
-  describe('Nx 17.x, Angular 16.x', () => {
+  describe('Nx 18.x, Angular 17.x', () => {
     it('creates a default analogjs app in the source directory', async () => {
       const analogAppName = 'analog';
       const { config, tree } = await setup({ analogAppName });
+      const { dependencies, devDependencies } = readJson(tree, 'package.json');
+
+      verifyCoreDependenciesNxV18_AngularV17_X(dependencies, devDependencies);
+
+      verifyConfig(config, analogAppName);
+
+      verifyHomePageExists(tree, analogAppName);
+
+      verifyEslint(tree, config, devDependencies);
+    });
+
+    it('creates a default standalone analogjs app in the source directory', async () => {
+      const analogAppName = 'analog';
+      const { config, tree } = await setup({ analogAppName }, '18.0.0', true);
+      const { dependencies, devDependencies } = readJson(tree, 'package.json');
+
+      verifyCoreDependenciesNxV18_AngularV17_X(dependencies, devDependencies);
+
+      verifyConfig(config, analogAppName, true);
+
+      verifyHomePageExists(tree, analogAppName, true);
+
+      verifyEslint(tree, config, devDependencies);
+    });
+
+    it('creates an analogjs app in the source directory with tailwind set up', async () => {
+      const analogAppName = 'tailwind-app';
+      const { config, tree } = await setup({
+        analogAppName,
+        addTailwind: true,
+      });
+      const { dependencies, devDependencies } = readJson(tree, 'package.json');
+
+      verifyCoreDependenciesNxV18_AngularV17_X(dependencies, devDependencies);
+
+      verifyConfig(config, analogAppName);
+
+      verifyHomePageExists(tree, analogAppName);
+
+      verifyTailwindIsSetUp(tree, devDependencies);
+    });
+
+    it('creates an analogjs app in the source directory with trpc set up', async () => {
+      const analogAppName = 'trpc-app';
+      const { config, tree } = await setup({ analogAppName, addTRPC: true });
+      const { dependencies, devDependencies } = readJson(tree, 'package.json');
+
+      verifyCoreDependenciesNxV18_AngularV17_X(dependencies, devDependencies);
+
+      verifyConfig(config, analogAppName);
+
+      verifyHomePageExists(tree, analogAppName);
+      verifyTrpcIsSetUp(tree, dependencies);
+    });
+  });
+
+  describe('Nx 17.x, Angular 16.x', () => {
+    it('creates a default analogjs app in the source directory', async () => {
+      const analogAppName = 'analog';
+      const { config, tree } = await setup({ analogAppName }, '17.0.0');
       const { dependencies, devDependencies } = readJson(tree, 'package.json');
 
       verifyCoreDependenciesNxV17_AngularV16_X(dependencies, devDependencies);
@@ -266,10 +355,13 @@ describe('nx-plugin generator', () => {
 
     it('creates an analogjs app in the source directory with tailwind set up', async () => {
       const analogAppName = 'tailwind-app';
-      const { config, tree } = await setup({
-        analogAppName,
-        addTailwind: true,
-      });
+      const { config, tree } = await setup(
+        {
+          analogAppName,
+          addTailwind: true,
+        },
+        '17.0.0'
+      );
       const { dependencies, devDependencies } = readJson(tree, 'package.json');
 
       verifyCoreDependenciesNxV17_AngularV16_X(dependencies, devDependencies);
@@ -283,7 +375,10 @@ describe('nx-plugin generator', () => {
 
     it('creates an analogjs app in the source directory with trpc set up', async () => {
       const analogAppName = 'trpc-app';
-      const { config, tree } = await setup({ analogAppName, addTRPC: true });
+      const { config, tree } = await setup(
+        { analogAppName, addTRPC: true },
+        '17.0.0'
+      );
       const { dependencies, devDependencies } = readJson(tree, 'package.json');
 
       verifyCoreDependenciesNxV17_AngularV16_X(dependencies, devDependencies);