+++ packages/vite-plugin-angular/src/lib/angular-vite-plugin.ts
@@ -16,12 +16,7 @@ import { jitPlugin } from './angular-jit-plugin';
 import { angularVitestPlugin } from './angular-vitest-plugin';
 
 import { createCompilerPlugin } from './compiler-plugin';
-import {
-  hasStyleUrls,
-  hasTemplateUrl,
-  StyleUrlsResolver,
-  TemplateUrlsResolver,
-} from './component-resolvers';
+import { StyleUrlsResolver, TemplateUrlsResolver } from './component-resolvers';
 import { augmentHostWithResources } from './host';
 import {
   angularApplicationPreset,
@@ -29,6 +24,7 @@ import {
   loadEsmModule,
   SourceFileCache,
 } from './utils/devkit';
+import { getFrontmatterMetadata } from './authoring/frontmatter';
 
 export interface PluginOptions {
   tsconfig?: string;
@@ -45,7 +41,11 @@ export interface PluginOptions {
     /**
      * Enable experimental support for Analog file extension
      */
-    supportAnalogFormat?: boolean;
+    supportAnalogFormat?:
+      | boolean
+      | {
+          include: string[];
+        };
   };
   supportedBrowsers?: string[];
   transformFilter?: (code: string, id: string) => boolean;
@@ -56,6 +56,8 @@ interface EmitFileResult {
   map?: string;
   dependencies: readonly string[];
   hash?: Uint8Array;
+  errors: string[];
+  warnings: string[];
 }
 type FileEmitter = (file: string) => Promise<EmitFileResult | undefined>;
 
@@ -64,7 +66,7 @@ type FileEmitter = (file: string) => Promise<EmitFileResult | undefined>;
  * Match .(c or m)ts, .ts extensions with an optional ? for query params
  * Ignore .tsx extensions
  */
-const TS_EXT_REGEX = /\.[cm]?(ts|analog)[^x]?\??/;
+const TS_EXT_REGEX = /\.[cm]?(ts|analog|ag)[^x]?\??/;
 
 export function angular(options?: PluginOptions): Plugin[] {
   /**
@@ -262,6 +264,13 @@ export function angular(options?: PluginOptions): Plugin[] {
           return;
         }
 
+        /**
+         * Skip transforming content files
+         */
+        if (id.includes('analog-content-list=true')) {
+          return;
+        }
+
         if (TS_EXT_REGEX.test(id)) {
           if (id.includes('.ts?')) {
             // Strip the query string off the ID
@@ -281,16 +290,8 @@ export function angular(options?: PluginOptions): Plugin[] {
             }
           }
 
-          let templateUrls: string[] = [];
-          let styleUrls: string[] = [];
-
-          if (hasTemplateUrl(code)) {
-            templateUrls = templateUrlsResolver.resolve(code, id);
-          }
-
-          if (hasStyleUrls(code)) {
-            styleUrls = styleUrlsResolver.resolve(code, id);
-          }
+          const templateUrls = templateUrlsResolver.resolve(code, id);
+          const styleUrls = styleUrlsResolver.resolve(code, id);
 
           if (watchMode) {
             for (const urlSet of [...templateUrls, ...styleUrls]) {
@@ -302,7 +303,18 @@ export function angular(options?: PluginOptions): Plugin[] {
             }
           }
 
-          const typescriptResult = fileEmitter && (await fileEmitter!(id));
+          const typescriptResult = await fileEmitter?.(id);
+
+          if (
+            typescriptResult?.warnings &&
+            typescriptResult?.warnings.length > 0
+          ) {
+            this.warn(`${typescriptResult.warnings.join('\n')}`);
+          }
+
+          if (typescriptResult?.errors && typescriptResult?.errors.length > 0) {
+            this.error(`${typescriptResult.errors.join('\n')}`);
+          }
 
           // return fileEmitter
           let data = typescriptResult?.content ?? '';
@@ -345,13 +357,18 @@ export function angular(options?: PluginOptions): Plugin[] {
           const useInputSourcemap = (!isProd ? undefined : false) as undefined;
 
           if (
-            id.includes('.analog') &&
+            (id.includes('.analog') || id.includes('.agx')) &&
             pluginOptions.supportAnalogFormat &&
             fileEmitter
           ) {
             sourceFileCache.invalidate([`${id}.ts`]);
             const ngFileResult = await fileEmitter!(`${id}.ts`);
             data = ngFileResult?.content || '';
+
+            if (id.includes('.agx')) {
+              const metadata = await getFrontmatterMetadata(code);
+              data += metadata;
+            }
           }
 
           if (!forceAsyncTransformation && !isProd) {
@@ -411,17 +428,34 @@ export function angular(options?: PluginOptions): Plugin[] {
   ].filter(Boolean) as Plugin[];
 
   function findAnalogFiles(config: UserConfig) {
-    if (!pluginOptions.supportAnalogFormat) {
+    const analogConfig = pluginOptions.supportAnalogFormat;
+    if (!analogConfig) {
       return [];
     }
 
+    let extraGlobs: string[] = [];
+
+    if (typeof analogConfig === 'object') {
+      if (analogConfig.include) {
+        extraGlobs = analogConfig.include;
+      }
+    }
+
     const fg = require('fast-glob');
-    const root = normalizePath(
+    const appRoot = normalizePath(
       path.resolve(pluginOptions.workspaceRoot, config.root || '.')
     );
+    const workspaceRoot = normalizePath(
+      path.resolve(pluginOptions.workspaceRoot)
+    );
+
+    const globs = [
+      `${appRoot}/**/*.{analog,agx}`,
+      ...extraGlobs.map((glob) => `${workspaceRoot}${glob}.{analog,agx}`),
+    ];
 
     return fg
-      .sync([`${root}/**/*.analog`], {
+      .sync(globs, {
         dot: true,
       })
       .map((file: string) => `${file}.ts`);
@@ -433,6 +467,7 @@ export function angular(options?: PluginOptions): Plugin[] {
       compilerCli.readConfiguration(pluginOptions.tsconfig, {
         suppressOutputPathCheck: true,
         outDir: undefined,
+        sourceMap: false,
         inlineSourceMap: !isProd,
         inlineSources: !isProd,
         declaration: false,
@@ -440,7 +475,11 @@ export function angular(options?: PluginOptions): Plugin[] {
         allowEmptyCodegenFiles: false,
         annotationsAs: 'decorators',
         enableResourceInlining: false,
+        noEmitOnError: false,
+        mapRoot: undefined,
+        sourceRoot: undefined,
         supportTestBed: false,
+        supportJitMode: false,
       });
 
     if (pluginOptions.supportAnalogFormat) {
@@ -477,7 +516,7 @@ export function angular(options?: PluginOptions): Plugin[] {
       | ts.BuilderProgram
       | ts.EmitAndSemanticDiagnosticsBuilderProgram;
     let typeScriptProgram: ts.Program;
-    let angularCompiler: any;
+    let angularCompiler: NgtscProgram['compiler'];
 
     if (!jit) {
       // Create the Angular specific program that contains the Angular compiler
@@ -541,9 +580,10 @@ export function angular(options?: PluginOptions): Plugin[] {
           afterDeclarations:
             pluginOptions.advanced.tsTransformers.afterDeclarations,
         },
-        jit ? {} : angularCompiler.prepareEmit().transformers
+        jit ? {} : angularCompiler!.prepareEmit().transformers
       ),
-      () => []
+      () => [],
+      angularCompiler!
     );
   }
 }
@@ -551,13 +591,26 @@ export function angular(options?: PluginOptions): Plugin[] {
 export function createFileEmitter(
   program: ts.BuilderProgram,
   transformers: ts.CustomTransformers = {},
-  onAfterEmit?: (sourceFile: ts.SourceFile) => void
+  onAfterEmit?: (sourceFile: ts.SourceFile) => void,
+  angularCompiler?: NgtscProgram['compiler']
 ): FileEmitter {
   return async (file: string) => {
     const sourceFile = program.getSourceFile(file);
     if (!sourceFile) {
       return undefined;
     }
+
+    const diagnostics = angularCompiler
+      ? angularCompiler.getDiagnosticsForFile(sourceFile, 1)
+      : [];
+
+    const errors = diagnostics
+      .filter((d) => d.category === ts.DiagnosticCategory.Error)
+      .map((d) => d.messageText);
+
+    const warnings = diagnostics
+      .filter((d) => d.category === ts.DiagnosticCategory.Warning)
+      .map((d) => d.messageText);
 
     let content: string | undefined;
     program.emit(
@@ -574,6 +627,6 @@ export function createFileEmitter(
 
     onAfterEmit?.(sourceFile);
 
-    return { content, dependencies: [] };
+    return { content, dependencies: [], errors, warnings };
   };
 }